Finished script. Added manager command.
