+++ README.md
@@ -1,6 +1,22 @@
-# Сайт - форум
-## Цвет
- color: #1f57ff
-## Этапы:
+# Сайт - форум (zenum)
 
-- Создание fish ✅
+## Создание
+
+Для создания базы данных используйте скрипт rebase.sh
+
+Поля для изменения:
+
+```shell
+$LOCAL_PATH - укажите путь к проекту
+$USER_COUNT - кол-во пользователей для рандомного добавления в БД
+$QUESTION_COUNT - кол-во вопросов
+$MAX_COUNT_ANSWERS_IN_QUESTION - максимальное кол-во ответов на вопросы
+$COUNT_RATING_PER_QUESTION - кол-во голосов для каждого вопроса (умножаем)
+```
+
+## Используемая палитра:
+
+![one.png](static/img/img_1.png)
+![two.png](static/img/img_2.png)
+![three.png](static/img/img_3.png)
+![four.png](static/img/img_4.png)
