+++ templates/questions.html
@@ -9,7 +9,12 @@
                     <div class="col-12">
                         <div class="d-flex flex-row">
                             <div class="mx-4 d-inline-block text-center">
-                                <img src="{% static 'img/test.png' %}" alt="test" class="img-question">
+                                {% if question.user.profile.profile_pic %}
+                                    <img src="{{ question.user.profile.profile_pic.url }}" alt="test"
+                                         class="profile-pic">
+                                {% else %}
+                                    <img src="{% static "img/nothing_to_show.svg" %}" alt="test" class="profile-pic">
+                                {% endif %}
                                 {% if user.user.is_authenticated %}
                                     <form class="pt-3">
                                         {% csrf_token %}
